Change Datatype to Long at all master repositories

+++ src/main/java/com/app/smartdrive/api/repositories/master/CarsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarsRepository extends JpaRepository<CarSeries, Integer> {
+public interface CarsRepository extends JpaRepository<CarSeries, Long> {
 }
